api/requirements.txt
langchain==0.0.194
langchain_serve==0.0.37
python-dotenv==1.0.0
